Refactoring pairwise 3d test suite
